requirements.txt
gunicorn
flask
tensorflow-cpu
numpy
IPython
pathlib
